refactor(core-utils): move itineraryToTransitive to transitive-overlay
BREAKING CHANGE: itineraryToTransitive function now in transitive-overlay package
